requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart
jinja2
pydantic[email]